chore: rename Header to SheetHeader
