+++ client/src/lib/components/character-sheet/header/SheetHeader.svelte
@@ -1,0 +1,254 @@
+<script lang="ts">
+	import Select from "$lib/components/Select.svelte";
+	import Char from "./Char.svelte";
+	import Info from "./Info.svelte";
+	import {
+		type Alignments2024Type,
+		type BackgroundTypes,
+		type ClassesType,
+		type RacesType,
+		VALID_ALIGNMENTS_2024,
+		VALID_BACKGROUNDS,
+		VALID_CLASSES,
+		VALID_RACES
+	} from "@types";
+
+	let charName = $state("");
+	let selectedClass: ClassesType | "" = $state("");
+	let selectedBackground: BackgroundTypes | "" = $state("");
+	let selectedRace: RacesType | "" = $state("");
+	let selectedAlignment: Alignments2024Type | "" = $state("");
+	let playerName = $state("");
+	let exp = $state(0);
+</script>
+
+<!-- MOBILE -->
+<div class="justify-center w-full md:hidden">
+	<Char bind:value={charName} />
+	<Info
+		classOptions={VALID_CLASSES}
+		backgroundOptions={VALID_BACKGROUNDS}
+		raceOptions={VALID_RACES}
+		alignmentOptions={VALID_ALIGNMENTS_2024}
+		bind:selectedAlignment
+		bind:selectedRace
+		bind:selectedClass
+		bind:selectedBackground
+		bind:playerName
+		bind:exp
+	/>
+</div>
+
+<!-- DESKTOP -->
+<div class="hidden md:block">
+	<svg
+		width="100%"
+		height="100%"
+		viewBox="0 0 1075 181"
+		fill="none"
+		xmlns="http://www.w3.org/2000/svg"
+	>
+		<path
+			d="M286 30.9903C284.442 30.0281 286.247 25.7399 288.724 24.5178C289.563 24.1039 290.925 23.761 291.75 23.756C293.755 23.743 296.401 25.42 297.053 27.1165C297.653 28.6775 297.342 31.4761 296.449 32.5513C295.554 33.6289 293.408 31.5281 293.802 29.9596C294.242 28.2076 293.167 26.6536 291.642 26.8366C290.242 27.0045 289.831 27.5593 289.12 30.2405L288.656 31.99L287.997 30.6154C287.219 28.9904 286.5 28.8245 286.5 30.27C286.5 30.8358 286.275 31.1602 286 30.9903Z"
+			fill="white"
+		/>
+		<path
+			d="M266.211 38.194C264.293 36.0762 265.825 31.6706 268.687 31.0707C269.815 30.8343 274.493 31.2022 275.125 31.5766C275.331 31.699 275.5 32.7042 275.5 33.8104C275.5 35.2249 275.315 35.7453 274.875 35.5628C272.667 34.6476 270.464 34.1463 269.788 34.4057C268.311 34.972 268.997 35.7098 271.739 36.507C275.07 37.4747 276.5 39.0732 276.5 41.8284C276.5 45.3878 273.022 47.2412 268 46.3574L265.75 45.9621L265.599 43.054L265.448 40.1464L267.099 41.376C269.031 42.8141 270.973 43.3709 271.952 42.7661C273.152 42.0248 272.163 40.8717 269.595 40.0189C268.232 39.5666 266.71 38.7453 266.211 38.194Z"
+			fill="white"
+		/>
+		<path
+			d="M407.216 34.4962C407.566 32.6288 408.302 31.8635 410.343 31.2427C411.774 30.8078 415.974 31.0972 416.586 31.6736C416.771 31.8475 417.084 32.8697 417.282 33.9448C417.611 35.7333 417.561 35.8662 416.696 35.5009C414.127 34.4162 411.992 34.0473 411.446 34.5931C410.681 35.3584 411.222 36.0652 412.871 36.4506C415.664 37.1033 417.064 37.9126 417.773 39.2832C420.06 43.7043 416.165 47.3836 410.25 46.3889C407.753 45.9691 407.75 45.9656 407.292 43.6033C406.739 40.7562 407.045 40.3238 408.798 41.4725C410.692 42.7126 412.697 43.1954 413.677 42.6471C414.911 41.9573 414.26 41.312 410.738 39.7325C407.321 38.1995 406.696 37.2693 407.216 34.4962Z"
+			fill="white"
+		/>
+		<path
+			d="M173 36.9344C173 33.3925 172.84 32.4788 172.125 31.936C171.368 31.3617 171.69 31.2697 174.5 31.2557L177.75 31.2392V36.8629V42.4867L179.292 42.6371C181.74 42.876 182.119 41.9578 181.923 36.2546L181.75 31.2392L186.986 31.2397L192.223 31.2402L195.361 34.3497L198.5 37.4592V35.128C198.5 33.2851 198.267 32.6333 197.389 32.0184C196.343 31.2857 196.507 31.2402 200.194 31.2402C203.992 31.2402 204.079 31.2667 203.059 32.1149C202.1 32.9127 201.996 33.577 201.879 39.666L201.75 46.3419L197.554 42.2902C195.246 40.0614 193.278 38.238 193.179 38.238C193.081 38.238 193 39.6261 193 41.3225C193 43.4914 193.245 44.6785 193.827 45.3213C194.622 46.199 194.508 46.235 190.952 46.2195C187.766 46.206 187.372 46.111 188.125 45.5397C188.877 44.9694 189 43.9912 189 38.6114C189 33.1426 188.891 32.3049 188.138 32.0164C186.576 31.4166 186 33.0981 186 38.2605C186 42.6596 185.893 43.2054 184.805 44.3786C183.397 45.8981 180.111 46.9053 178.143 46.4204C174.223 45.4552 173 43.1994 173 36.9344Z"
+			fill="white"
+		/>
+		<path
+			d="M206.669 44.9389C202.228 41.3685 202.721 35.2879 207.706 32.1569C208.929 31.3891 209.947 31.2217 212.598 31.3532C214.439 31.4451 216.09 31.6631 216.265 31.8385C216.791 32.3638 217.167 36.2401 216.691 36.2211C216.449 36.2111 215.688 35.8817 215 35.4889C211.798 33.6589 208 35.4849 208 38.8538C208 39.7675 208.331 40.9871 208.736 41.5645C209.487 42.6371 211.663 43.3344 212.844 42.8815C213.931 42.4642 213.627 39.8835 212.388 39.0163C211.334 38.2775 211.442 38.238 214.513 38.238H217.75V42.0088C217.75 45.6032 217.691 45.7956 216.5 46.1325C215.812 46.3265 213.751 46.4844 211.919 46.4834C208.947 46.4819 208.382 46.3155 206.669 44.9389Z"
+			fill="white"
+		/>
+		<path
+			d="M220.5 38.7378C220.5 34.1638 220.34 32.8917 219.672 32.1544C218.875 31.2737 219.057 31.2402 224.622 31.2402H230.4L230.578 33.4895C230.748 35.6323 230.421 36.3241 229.75 35.2389C229.58 34.964 228.442 34.7391 227.22 34.7391C225.333 34.7391 225 34.889 225 35.7388C225 36.5905 225.334 36.7385 227.25 36.7385H229.5V38.6164V40.4943L227.25 40.3158C225.114 40.1459 225 40.2029 225 41.4365C225 42.6166 225.182 42.7366 226.967 42.7366C228.049 42.7366 229.538 42.5072 230.275 42.2272C231.611 41.7194 231.614 41.7229 231.262 43.5978C231.068 44.632 230.804 45.6487 230.675 45.8566C230.546 46.065 227.832 46.2355 224.643 46.2355C219.058 46.2355 218.875 46.202 219.672 45.3213C220.34 44.584 220.5 43.3119 220.5 38.7378Z"
+			fill="white"
+		/>
+		<path
+			d="M250.5 38.8738C250.5 34.1733 250.342 32.8942 249.672 32.1544C248.882 31.2817 248.946 31.2402 251.065 31.2402C253.093 31.2402 253.519 31.4776 256 33.9893C257.493 35.5014 258.893 36.7385 259.108 36.7385C259.849 36.7385 259.5 33.0681 258.672 32.1544C257.878 31.2767 257.992 31.2407 261.547 31.2562C264.734 31.2697 265.128 31.3647 264.375 31.936C263.617 32.5113 263.5 33.511 263.5 39.4176C263.5 43.1674 263.326 46.2355 263.113 46.2355C262.9 46.2355 260.948 44.4475 258.775 42.2617C256.601 40.0764 254.638 38.4024 254.411 38.5424C254.185 38.6824 254 40.0119 254 41.497C254 43.6308 254.23 44.4111 255.095 45.2163L256.189 46.2355H252.233C248.511 46.2355 248.343 46.1895 249.388 45.4572C250.428 44.7295 250.5 44.3066 250.5 38.8738Z"
+			fill="white"
+		/>
+		<path
+			d="M362.486 45.4432C357.486 42.5246 357.643 34.5387 362.75 32.0065C364.479 31.1497 369.608 30.9953 371.033 31.7575C371.741 32.1364 372 32.8092 372 34.2697V36.2641L370.525 35.5019C369.713 35.0825 368.155 34.7391 367.061 34.7391C364.332 34.7391 362.789 36.615 363.169 39.4701C363.479 41.7874 364.547 42.7366 366.845 42.7366C368.391 42.7366 368.5 42.6261 368.5 41.0481C368.5 39.8675 368.161 39.1187 367.375 38.5589C366.301 37.7951 366.398 37.7636 369.5 37.8731L372.75 37.9881L372.896 41.8354L373.041 45.6827L371.146 46.0735C367.814 46.7603 364.304 46.5049 362.486 45.4432Z"
+			fill="white"
+		/>
+		<path
+			d="M392 38.7378C392 33.4785 391.875 32.5048 391.125 31.936C390.359 31.3552 390.562 31.2702 392.737 31.2562C395.124 31.2407 395.349 31.3646 398.362 34.3497L401.5 37.4592V35.128C401.5 33.2851 401.267 32.6333 400.388 32.0184C399.342 31.2857 399.507 31.2402 403.194 31.2402C406.992 31.2402 407.079 31.2667 406.06 32.1149C405.102 32.9117 404.996 33.5785 404.879 39.6021L404.75 46.2145L400.5 42.2637L396.25 38.313L396.099 41.332C395.985 43.6298 396.153 44.5765 396.802 45.2933C397.631 46.2085 397.549 46.235 393.953 46.2195C390.766 46.206 390.372 46.111 391.125 45.5397C391.875 44.9709 392 43.9972 392 38.7378Z"
+			fill="white"
+		/>
+		<path
+			d="M302.266 35.4254C302.629 35.5653 303.164 35.3934 303.454 35.044C304.136 34.2223 308.878 34.3677 310.522 35.2609C311.744 35.9247 311.739 35.9312 309.517 36.5825C308.289 36.9424 306.913 37.1194 306.458 36.9749C304.982 36.5065 303.082 37.6477 301.903 39.7105C300.685 41.8434 299.984 42.1733 299.054 41.0521C298.606 40.5123 298.911 39.9824 300.494 38.5529C301.599 37.5547 302.351 36.7385 302.166 36.7385C301.98 36.7385 301.417 37.1109 300.914 37.5657C299.786 38.5864 299.721 37.8596 300.802 36.3161C301.243 35.6868 301.902 35.2859 302.266 35.4254Z"
+			fill="white"
+		/>
+		<path
+			d="M284.147 49.2775C282.384 49.6989 282.033 40.6342 283.74 38.7488C284.968 37.3923 285.153 37.4702 284.893 39.2377C284.636 40.9896 284.997 41.1226 286.25 39.7375C287.317 38.5589 288.089 38.4239 287.681 39.4876C287.362 40.3183 288.261 40.5288 288.75 39.7375C289.498 38.5269 289.95 39.3846 289.263 40.7122C286.68 45.7052 290.827 49.9843 294.596 46.216C296.039 44.7734 296.176 44.736 297.308 45.4777C298.884 46.5099 298.824 46.8738 296.695 49.2295C293.649 52.6 289.206 53.1043 286.036 50.4382C285.215 49.7474 284.365 49.225 284.147 49.2775Z"
+			fill="white"
+		/>
+		<path
+			d="M294.25 40.2374C293.838 39.895 293.5 39.4386 293.5 39.2232C293.5 38.4939 295.528 39.3561 296.034 40.2999C296.31 40.8152 296.824 41.2371 297.177 41.2371C297.53 41.2371 299.006 42.1218 300.456 43.2034C303.024 45.1183 304.5 45.3743 304.5 43.9032C304.5 43.5038 304.725 43.3164 305 43.4863C305.275 43.6563 305.5 44.5855 305.5 45.5517C305.5 46.9133 305.685 47.2722 306.322 47.1467C307.665 46.8828 308.343 45.3992 307.637 44.2691C307.066 43.3549 307.119 43.3299 308.555 43.8302C311.77 44.9509 311.728 48.9846 308.49 50.0643L306.75 50.6441L308 51.1375L309.25 51.6308L307.842 52.1821C307.068 52.4855 306.03 52.7334 305.535 52.7334C304.35 52.7334 302 50.8625 302 49.9183C302 49.51 301.756 49.3265 301.457 49.5115C301.158 49.6959 301.024 50.2717 301.159 50.7906C301.295 51.3094 301.211 51.7338 300.971 51.7338C300.24 51.7338 299.5 49.8958 299.5 48.0779C299.5 46.6434 299.125 46.0485 297.25 44.513C296.013 43.4988 295 42.2622 295 41.7644C295 41.267 294.662 40.5798 294.25 40.2374Z"
+			fill="white"
+		/>
+		<path
+			fill-rule="evenodd"
+			clip-rule="evenodd"
+			d="M110.663 16.6628C110.368 11.4134 111.219 6.32851 113.051 2.38925L114.163 0L114.557 3.17101C115.394 9.92189 118.761 14.8663 130.268 26.2418C134.579 30.5029 138.795 34.945 139.637 36.1127C141.664 38.9238 141.991 38.4264 140.186 35.2774C138.098 31.6336 133.866 26.7221 128.908 22.189C122.83 16.6313 122.321 15.6941 122.61 10.5892C122.747 8.17843 123.214 5.67672 123.747 4.49859L124.651 2.49922L125.087 5.22686C126.077 11.4354 128.806 15.4836 139.25 26.2418C143.387 30.5029 147.507 35.148 148.406 36.5645C150.662 40.1184 152 44.1906 152 47.5061C152 50.0533 152.156 50.4477 153.75 51.9512C154.712 52.8589 155.5 54.081 155.5 54.6668C155.5 56.2568 156.278 55.9799 159.021 53.4147L161.5 51.097V53.0598C161.5 55.5386 160.304 60.1741 159.012 62.7068L158.019 64.6522L161.135 64.9501C162.848 65.1141 170.438 65.5864 178 65.9998C185.562 66.4131 196.363 67.0929 202 67.5103C207.637 67.9277 215.738 68.4835 220 68.7454C224.262 69.0068 230.675 69.4492 234.25 69.7281C244.792 70.5499 254.201 71.2121 262.485 71.715C266.755 71.9744 275.537 72.5257 282 72.9401C311.079 74.8055 328.174 75.3833 354.75 75.3988C380.121 75.4138 387.825 75.1534 405.25 73.6919C410.812 73.2255 432.174 70.8612 438 70L437.5 72.5L424.22 74.1912C383.549 79.8194 325.356 79.3571 239.75 72.7252C236.175 72.4483 226.725 71.779 218.75 71.2371C210.775 70.6958 200.987 70.016 197 69.7271C190.539 69.2588 186.395 69.0143 167.25 67.9757C164.5 67.8267 160.9 67.6058 159.25 67.4848C156.438 67.2789 156.13 67.3673 154.339 68.9014L152.429 70.5374L164.839 71.1737C178.848 71.8919 207.476 73.8453 220.25 74.9545C255.069 77.978 271.421 79.1407 302.5 80.8036C330.107 82.2807 384.972 82.4256 401.25 81.0651C412.851 80.0952 420.947 79.0467 427 78C427.5 78 438.183 76.5581 438.5 76.5V78.5L425.5 80C422.125 80.5438 417.225 80.9246 414.75 81.1955C387.443 84.1856 337.153 84.1941 286 81.217C278.34 80.7711 274.545 80.5212 255.75 79.2256C251.9 78.9602 242 78.1725 233.75 77.4757C189.819 73.7643 166.393 72.2888 138.125 71.4526C130.459 71.2256 127.5 71.2931 127.5 71.695C127.5 72.4543 129.99 73.3455 131.374 73.0811C133.005 72.7692 132.769 73.4579 130.451 75.7757C127.984 78.2419 123.145 80.6042 119.738 81.0061L117.25 81.3L118.505 80.1373C119.196 79.498 120.399 78.138 121.179 77.1153L122.597 75.2549L119.195 75.5728L115.792 75.8902L117.396 73.9448C118.278 72.8751 119 71.826 119 71.6135C119 71.2706 95.052 71.2326 83.25 71.557C71.2495 71.8869 47.8395 73.5429 37.665 74.7815C33.389 75.3024 33.0585 75.4243 32.7645 76.596L32.449 77.8521L31.3175 76.7894C30.007 75.5588 28.0975 75.3908 27.683 76.4695C27.526 76.8789 27.642 77.9135 27.9405 78.7693C28.4245 80.1568 28.3895 80.2888 27.6165 79.9899C27.14 79.8059 26.162 79.5365 25.4435 79.3911L24.137 79.1277L24.0385 106.667C23.984 121.813 24.052 134.387 24.189 134.609C24.3265 134.831 27.4335 134.619 31.094 134.137C71.1425 128.871 128.428 128.274 190.25 132.481C200.182 133.157 213.928 134.238 234.25 135.943C265.49 138.564 310.261 141.178 324.277 141.199C328.447 141.205 328.99 141.088 331.151 139.718C337.002 136.008 351.87 129.897 360.25 127.756C374.102 124.217 387.61 124.309 396.707 128.005C400.017 129.35 400.169 129.388 399.793 128.779C399.27 127.933 392.7 125.103 389.009 124.134C383.589 122.711 372.58 122.339 367.032 123.391C364.712 123.831 362.733 124.111 362.636 124.014C362.293 123.671 367.98 120.179 371.348 118.665C375.994 116.575 383.175 114.712 386.5 114.735L389.25 114.754L387.75 115.546C385.954 116.494 384.5 117.715 384.5 118.273C384.5 119.358 386.548 119.681 390.33 119.195C396.159 118.444 403.405 118.544 405.75 119.406L407.75 120.142L405.5 120.227C402.478 120.341 400.953 120.814 400.309 121.834C399.524 123.079 400.256 123.566 404.642 124.714C409.262 125.924 411.362 126.846 415 129.263C417.803 131.125 419.841 133.137 421.96 136.137C423.112 137.768 423.219 137.807 425.698 137.482C427.101 137.299 429.207 136.91 430.375 136.619V138.276H428.5C428.035 138.372 427.617 138.456 427.242 138.531C425.171 138.948 424.418 139.1 424.315 139.446C424.256 139.643 424.409 139.904 424.649 140.315C424.725 140.445 424.81 140.59 424.9 140.752C425.733 142.262 425.807 142.286 428.668 142.002H430.375L429.875 144.464C428.431 144.659 427.192 144.87 427.12 144.933C427.048 144.996 427.329 146.012 427.745 147.189C428.932 150.552 428.763 151.02 426.875 149.596C425.981 148.921 424.19 147.753 422.893 147C420.617 145.678 420.411 145.648 416.893 146.148C405.59 147.753 390.098 148.662 367.505 149.048L348.76 149.368L345.418 151.909C340.735 155.472 331.72 160.867 328.786 161.863C320.55 164.658 310.51 160.376 302.815 150.784L300.444 147.829L295.597 147.499C292.931 147.318 287.375 146.958 283.25 146.7C255.212 144.947 243.493 144.167 241.75 143.941C240.65 143.799 233.113 143.236 225 142.691C216.887 142.145 207.213 141.474 203.5 141.198C199.787 140.922 192.925 140.48 188.25 140.216C183.575 139.952 176.15 139.508 171.75 139.23C141.183 137.297 92.966 136.393 76.14 137.436L70.03 137.815L69.89 143.828L69.75 149.84L68 150.022C67.0375 150.122 62.65 150.67 58.25 151.24C53.85 151.81 46.9875 152.697 43 153.211C39.0125 153.725 33.05 154.503 29.75 154.94C19.395 156.31 8.8855 157.708 5.25 158.198C3.325 158.457 1.3565 158.676 0.875 158.685C0.0495 158.699 0 156.625 0 121.992V85.2847L2.125 85.0083C3.294 84.8564 7.531 84.2805 11.541 83.7282C15.551 83.1759 19.011 82.724 19.23 82.724C19.4495 82.724 18.694 81.8808 17.552 80.8496L15.4755 78.9752L14.9115 73.6614C14.601 70.7388 14.469 68.0396 14.618 67.6633C14.7665 67.2869 15.5575 66.5911 16.3745 66.1172L17.861 65.255L18.4535 71.3656C18.962 76.6035 19.2015 77.6071 20.1335 78.3949C21.582 79.6195 21.707 79.0357 21.047 74.1317C20.36 69.0278 20.356 65.5179 21.035 63.0737C21.439 61.6192 22.1125 60.8204 23.88 59.6998C25.1535 58.8925 26.3415 58.2317 26.52 58.2317C26.6985 58.2317 27.498 58.8025 28.297 59.5008C29.588 60.628 29.656 60.8239 28.9095 61.2558C28.422 61.5377 27.8195 62.9417 27.4755 64.5977C27.1495 66.1682 26.7725 67.7402 26.638 68.0906C26.4775 68.508 26.847 68.7284 27.707 68.7284C29.128 68.7284 29.574 68.0321 30.4965 64.3758C30.7885 63.2192 31.2035 62.164 31.419 62.0305C31.634 61.8976 32.717 62.019 33.825 62.2999C35.521 62.7303 36.0135 63.1687 36.9365 65.0671C37.723 66.6861 38.2515 67.2394 38.8075 67.026C39.2335 66.8625 39.7395 66.729 39.9325 66.729C40.125 66.729 40.067 65.167 39.803 63.2581C39.4325 60.5795 39.508 58.6681 40.1335 54.8858C41.224 48.2903 42.1535 45.5877 45.8505 38.259C49.202 31.6146 49.34 30.203 46.4785 31.845C45.665 32.3114 45 32.5673 45 32.4128C45 32.2589 45.6655 30.7513 46.479 29.0624C49.2155 23.3817 54.7705 17.8519 60.8125 14.7949C63.754 13.3063 73.8875 9.74694 75.183 9.74694C75.5275 9.74694 76.2455 9.40955 76.7795 8.99668C77.732 8.26091 77.728 8.22792 76.5695 7.24723C75.9205 6.6979 74.908 6.24104 74.3195 6.23304C72.119 6.20155 74.9265 5.20837 80.1315 4.17669C89.0785 2.40375 98.132 3.40193 105 6.91833L108.25 8.5823L108.561 14.5379C108.732 17.8134 108.814 20.5605 108.745 20.6425C108.675 20.724 106.32 19.2994 103.51 17.4765C98.898 14.485 98.173 13.8012 96.075 10.4667C94.796 8.43435 93.525 6.84035 93.25 6.92383C92.975 7.0073 92.2475 7.958 91.633 9.03617C90.4395 11.132 90.073 16.057 90.9415 18.3402L91.383 19.5009L84.7175 19.4974C80.6545 19.4954 77.843 19.7023 77.517 20.0277C77.1375 20.4061 77.667 21.0529 79.34 22.254C85.243 26.4917 86.9975 28.8509 85.8955 31.0682C85.187 32.4933 84.4445 32.9352 76 36.9504C72.0445 38.8313 69.506 40.4078 67.866 42.0023C65.533 44.2706 63.677 47.0787 64.1785 47.5806C64.3175 47.719 65.5825 47.5856 66.99 47.2842C71.794 46.254 77.4535 46.7443 79.698 48.3848C80.5555 49.0116 80.539 49.1326 79.359 50.871C77.349 53.8321 77.144 56.9371 78.629 61.9286L79.2085 63.876L86.229 63.555C90.0905 63.3781 99.1565 63.2332 106.375 63.2322C115.712 63.2307 119.5 63.0687 119.5 62.6713C119.5 61.5607 115.688 58.3307 113.738 57.7894C110.94 57.0126 105.787 57.0986 103.734 57.9563C102.78 58.3547 102 58.5056 102 58.2917C102 58.0783 102.696 56.8536 103.546 55.5706C105.329 52.8789 109.02 49.467 111.375 48.3323C112.269 47.902 113 47.3961 113 47.2087C113 47.0207 111.594 46.4869 109.875 46.0226L109.524 45.9277C106.98 45.241 105.896 44.9484 105.876 44.5974C105.862 44.3368 106.435 44.0439 107.432 43.5337L107.617 43.4389C110.058 42.1873 117.101 40.2949 119.75 40.1784C120.916 40.1269 120.749 40.0279 119 39.734C113.346 38.7838 104.145 41.0281 99.3565 44.524C96.746 46.4304 96.6885 47.2142 99.1545 47.2737C100.032 47.2947 101.65 47.5886 102.75 47.926L104.75 48.5403L103.25 48.9246C98.36 50.1778 89.5775 54.6838 84.875 58.3527C82.7245 60.0302 82.6965 59.9402 83.7695 54.7983C84.3725 51.9097 84.416 50.7451 83.973 49.402C82.7175 45.6002 78.6405 43.8472 72.4005 44.4271C70.136 44.6375 69.04 44.5755 69.242 44.2491C69.832 43.2944 73.785 40.8727 80.25 37.5052C83.825 35.6428 87.256 33.6574 87.875 33.0926C90.128 31.0358 89.146 28.2201 84.6665 23.894L81.3715 20.711L83.3105 20.9579C84.3775 21.0934 86.7715 21.3323 88.6305 21.4883C90.49 21.6442 92.571 21.9851 93.2555 22.245C93.3349 22.2752 93.4099 22.3041 93.4806 22.3313C93.9684 22.5193 94.2527 22.6288 94.3623 22.5347C94.5155 22.4034 94.3282 21.875 93.8793 20.6088L93.8787 20.6071L93.8699 20.5822C93.8262 20.4589 93.78 20.3286 93.7315 20.1912C92.8385 17.6615 92.573 10.9736 93.335 10.2118C93.696 9.85091 94.0605 10.1863 94.524 11.304C95.709 14.1636 100.492 18.3807 107.5 22.7444C111.489 25.2276 115.509 28.1956 117.327 29.9986C119.267 31.9235 120.244 32.6328 119.973 31.9195C119.067 29.5377 116.98 26.8066 114.01 24.1204L110.925 21.3298L110.663 16.6628ZM21.5 116.463V143.205H22.743C23.4265 143.205 26.97 143.66 30.618 144.217C40.278 145.691 40.6195 145.751 40.42 145.951C40.189 146.182 4.9235 151.203 3.532 151.203C2.501 151.203 2.5 151.179 2.5 121.462C2.5 93.4822 2.552 91.7192 3.375 91.6907C3.8565 91.6737 7.2875 91.2499 11 90.7485C14.7125 90.2472 18.594 89.8113 19.625 89.7793L21.5 89.7218V116.463ZM20.8045 88.2223C21.244 88.2223 21.5 87.6705 21.5 86.7228C21.5 86.2805 21.5001 85.945 21.3457 85.7117C20.8602 84.9782 18.8476 85.2553 10.5036 86.4044L10.5 86.4049L2.75 87.4725L2.592 88.8471C2.3945 90.5626 2.598 90.5701 12.6795 89.2185C16.7655 88.6707 20.422 88.2223 20.8045 88.2223ZM67.3945 142.58C67.261 146.623 67.1745 146.944 66.25 146.825C65.7 146.754 60.6375 146.019 55 145.191C49.3625 144.363 42.094 143.32 38.8485 142.872C35.6025 142.425 33.1275 141.898 33.3485 141.703C33.783 141.317 39.0405 140.586 47.75 139.703C58.2635 138.635 61.7665 138.339 64.6445 138.273L67.539 138.207L67.3945 142.58ZM57.884 148.421C57.7945 148.331 55.382 147.909 52.5225 147.481C46.5995 146.595 45.838 146.594 40 147.462C36.1145 148.039 28.777 149.104 10.554 151.735C6.8715 152.267 3.5525 152.702 3.179 152.702C2.758 152.702 2.5 153.271 2.5 154.202C2.5 155.417 2.71 155.701 3.6095 155.701C4.7565 155.701 53.474 149.406 56.3985 148.88C57.305 148.717 57.9735 148.51 57.884 148.421ZM135.909 41.1126C137.928 42.6061 140.5 46.0585 140.5 47.2752C140.5 48.0134 140.26 48.0964 139.091 47.761C138.316 47.5391 137.184 47.0482 136.576 46.6713C134.63 45.4642 131.737 38.4649 133.625 39.5321C133.969 39.726 134.996 40.4373 135.909 41.1126ZM139.043 58.0413C140.404 55.0447 140.285 53.5322 138.625 52.7264C135.949 51.4279 123.032 51.7518 119.4 53.2083C118.333 53.6362 118.366 53.6722 119.85 53.7016C121.357 53.7311 126.5 54.6818 126.5 54.9303C126.5 54.9932 126.016 56.0429 125.426 57.263C124.835 58.4831 124.247 60.401 124.12 61.5247L123.888 63.568L133.319 63.8974C138.506 64.0789 144.528 64.2278 146.702 64.2283C150.521 64.2298 150.627 64.1983 149.815 63.3021C148.759 62.1345 144.621 60.7804 140.934 60.395L138.107 60.1001L139.043 58.0413ZM149.375 67.03C151.339 67.2704 151.5 67.4023 151.5 68.7749C151.5 69.5912 151.219 70.3559 150.875 70.4744C150.532 70.5928 145.075 70.5294 138.75 70.3339L127.25 69.978L127.381 68.7284C127.39 68.638 127.399 68.5507 127.408 68.4664C127.474 67.8055 127.521 67.3329 127.762 67.0016C128.368 66.17 130.196 66.2284 136.61 66.4334L137.622 66.4656C142.917 66.6326 148.207 66.887 149.375 67.03ZM120.191 68.1486C120.365 67.2799 120.449 66.5221 120.378 66.4646C119.902 66.0753 97.7255 65.9288 85.5 66.2337C68.3365 66.6621 59.307 67.2834 41.072 69.2917L35.3935 69.9171L34.607 71.566C34.1745 72.4727 33.9405 73.334 34.0865 73.4804C34.233 73.6264 38.8295 73.3055 44.3015 72.7667C61.6675 71.0567 91.0815 69.7281 111.562 69.7281H119.875L120.191 68.1486ZM29.25 72.9771C29.25 74.2412 29.0345 74.5016 27.875 74.6346C26.5985 74.7815 26.5 74.6631 26.5 72.9771C26.5 71.2911 26.5985 71.1727 27.875 71.3196C29.0345 71.4526 29.25 71.713 29.25 72.9771ZM401.387 137.418C398.459 134.273 395.052 132.379 390.64 131.444C386.579 130.584 385.595 130.587 381.95 131.472C376.736 132.737 361.711 139.358 359.775 141.243C358.843 142.15 386.106 141.5 400 140.284L403.75 139.956L401.387 137.418ZM407.25 141.934C412.09 142.541 415.198 143.36 414.141 143.751C411.464 144.741 387.58 146.155 367.75 146.498L352.25 146.765L354.208 144.985C356.113 143.254 356.276 143.205 360.208 143.204C367.035 143.202 395.012 142.157 399.25 141.746C401.45 141.533 403.588 141.417 404 141.49C404.412 141.562 405.875 141.763 407.25 141.934ZM185.25 133.481C179.2 133.069 169.3 132.486 163.25 132.186C152.886 131.672 89.823 131.143 82.25 131.507C61.2585 132.515 55.022 132.926 44.25 134.012C37.375 134.704 30.0625 135.538 28 135.864L24.25 136.457L24.0905 138.403L23.931 140.348L25.8405 140.02C37.6205 137.994 50.191 136.606 67.5 135.419C90.0065 133.876 137.606 134.459 172.75 136.709C177.15 136.99 187.725 137.665 196.25 138.208C204.775 138.75 214.787 139.423 218.5 139.702C235.063 140.945 265.848 143.084 276 143.696C314.013 145.989 316.229 146.04 320.319 144.72C322.206 144.111 324.071 143.434 324.462 143.215C324.857 142.995 321.848 142.675 317.712 142.497C313.608 142.32 303.725 141.835 295.75 141.42C267.407 139.944 267.169 139.928 233.25 137.213C229.675 136.927 223.15 136.359 218.75 135.951C214.35 135.542 207.487 134.989 203.5 134.719C199.513 134.45 191.3 133.893 185.25 133.481Z"
+			fill="white"
+		/>
+		<path
+			fill-rule="evenodd"
+			clip-rule="evenodd"
+			d="M236.121 32.2374C233.447 33.8669 232.413 36.0002 232.634 39.4351C232.804 42.0673 233.016 42.5571 234.622 44.0287C236.653 45.8891 239.783 46.9273 241.857 46.4289C246.345 45.3498 248.693 42.6681 248.693 38.6209C248.693 35.9277 247.655 34.1678 245.005 32.3648C242.812 30.8728 238.462 30.8103 236.121 32.2374ZM240.541 42.7366C244.094 42.7366 245.804 38.4974 243.272 35.9662C241.8 34.4942 239.578 34.3607 237.987 35.6488C237.145 36.3301 236.863 37.0844 236.863 38.6579C236.863 41.4235 238.048 42.7366 240.541 42.7366Z"
+			fill="white"
+		/>
+		<path
+			fill-rule="evenodd"
+			clip-rule="evenodd"
+			d="M290.085 31.3652C290.332 33.525 290.124 33.6239 288.318 32.2034C287.226 31.3447 287 31.3022 287 31.9535C287 32.3858 286.635 32.7397 286.189 32.7397C285.5 32.7397 285.479 32.8612 286.045 33.5435C286.612 34.2268 286.569 34.4522 285.76 35.043C285.237 35.4259 284.627 35.7388 284.405 35.7388C284.182 35.7388 284 35.9637 284 36.2386C284 36.5135 284.338 36.7385 284.75 36.7385C285.167 36.7385 285.5 37.1828 285.5 37.7382C285.5 39.1272 286.172 38.9888 287.923 37.2383C288.748 36.4136 289.533 35.7388 289.668 35.7388C290.012 35.7388 289.766 37.1388 289.348 37.5572C288.654 38.2505 289.026 38.7378 290.25 38.7378C291.345 38.7378 291.5 38.9543 291.5 40.4873C291.5 42.2907 291.762 42.5306 292.927 41.7929C293.446 41.4635 293.475 41.1251 293.052 40.3338C291.627 37.6727 293.491 36.7625 295.863 38.9608C296.589 39.6336 297.366 40.0704 297.591 39.931C298.145 39.5896 298.113 37.7382 297.554 37.7382C297.309 37.7382 296.969 37.1823 296.798 36.5025C296.628 35.8233 295.59 34.3517 294.494 33.2336C293.397 32.1149 292.5 30.9708 292.5 30.6909C292.5 30.4105 292.272 30.323 291.992 30.4954C291.713 30.6684 291.617 31.0752 291.78 31.3996C291.943 31.7245 291.58 31.4276 290.974 30.7404L289.871 29.4907L290.085 31.3652ZM294.785 37.2383C295.38 37.2383 295.424 37.0319 295 36.2386C294.706 35.6888 294.127 35.2389 293.715 35.2389C293.12 35.2389 293.076 35.4454 293.5 36.2386C293.794 36.7885 294.373 37.2383 294.785 37.2383Z"
+			fill="white"
+		/>
+		<path
+			fill-rule="evenodd"
+			clip-rule="evenodd"
+			d="M155 30.7054C155 30.9803 155.675 31.974 156.5 32.9132C157.948 34.5622 158 34.7901 158 39.5141C158 43.3659 157.824 44.602 157.173 45.3213C156.376 46.201 156.531 46.2355 161.298 46.2335C165.357 46.232 166.544 46.0525 167.88 45.2383C172.882 42.1893 172.66 34.7126 167.49 32.1159C166.228 31.4816 164.54 31.2417 161.342 31.2412C158.918 31.2407 156.499 31.0073 155.967 30.7229C155.333 30.384 155 30.378 155 30.7054ZM166.027 41.8154C167.71 40.4528 167.728 37.0369 166.058 35.6853C164.698 34.5846 162.584 34.4372 162.208 35.4164C162.065 35.7893 162.017 37.5322 162.1 39.2902C162.242 42.3077 162.322 42.4952 163.539 42.6366C164.264 42.7211 165.352 42.3617 166.027 41.8154Z"
+			fill="white"
+		/>
+		<path
+			fill-rule="evenodd"
+			clip-rule="evenodd"
+			d="M314 31.0338C314 31.0717 314.675 31.952 315.5 32.9896C316.926 34.7826 317 35.109 317 39.5566C317 43.5698 316.858 44.3786 316 45.2358L315 46.2355L319.875 46.2345C327.238 46.2335 329.954 44.3971 330.376 39.1332C330.551 36.9634 330.378 36.1746 329.413 34.7391C327.63 32.0874 325.577 31.3522 319.375 31.1447C316.418 31.0458 314 30.9958 314 31.0338ZM326 38.7378C326 37.4048 325.666 36.4051 325 35.7388C324.445 35.1835 323.334 34.7391 322.5 34.7391H321V38.7378V42.7366H322.5C324.638 42.7366 326 41.1801 326 38.7378Z"
+			fill="white"
+		/>
+		<path
+			fill-rule="evenodd"
+			clip-rule="evenodd"
+			d="M332 38.7378C332 34.1638 331.84 32.8917 331.172 32.1544C330.377 31.2752 330.513 31.2402 334.698 31.2402C339.225 31.2402 341.593 31.9585 342.19 33.512C342.777 35.042 342.499 37.1298 341.551 38.2995C340.607 39.4646 340.61 39.4826 342.092 41.354C342.913 42.3892 343.748 43.2364 343.948 43.2364C344.592 43.2364 348.597 33.1626 348.341 32.1844C348.122 31.3457 348.392 31.2402 350.758 31.2402H353.421L356.01 37.8631C357.492 41.6529 359.06 44.8554 359.675 45.3498C360.718 46.188 360.645 46.2135 357.194 46.2245C353.669 46.2355 353.641 46.225 353.878 44.9859C354.11 43.7743 354.032 43.7363 351.338 43.7363C348.411 43.7363 347.515 44.2991 348.518 45.5077C349.039 46.1345 348.576 46.2365 345.186 46.2425L341.25 46.249L338.645 43.3684L336.039 40.4873L336.02 42.4472C336.008 43.5938 336.344 44.7869 336.828 45.3213C337.622 46.199 337.509 46.2355 334 46.2355C330.491 46.2355 330.378 46.199 331.172 45.3213C331.84 44.584 332 43.3119 332 38.7378ZM338.345 35.057C338.731 36.0632 337.927 37.4322 336.822 37.6497C336.183 37.7756 336 37.4143 336 36.0252C336 34.5631 336.185 34.2393 337.016 34.2393C337.575 34.2393 338.172 34.6071 338.345 35.057ZM351.284 40.2374C352.384 40.2374 352.518 40.0764 352.223 39.1127C351.477 36.6755 351.102 36.4051 350.535 37.8956C349.759 39.937 349.855 40.2374 351.284 40.2374Z"
+			fill="white"
+		/>
+		<path
+			fill-rule="evenodd"
+			clip-rule="evenodd"
+			d="M377.655 32.2499C374.129 34.5706 373.087 39.849 375.488 43.2204C378.581 47.5616 385.863 47.5471 389.109 43.1929C391.366 40.1659 390.655 35.171 387.639 32.8717C385.182 30.9978 380.042 30.6779 377.655 32.2499ZM386 38.7573C386 37.3643 385.675 36.4755 384.905 35.7583C382.351 33.379 378.5 35.3609 378.5 39.0552C378.5 41.283 379.98 42.7366 382.25 42.7366C384.644 42.7366 386 41.2975 386 38.7573Z"
+			fill="white"
+		/>
+		<path
+			fill-rule="evenodd"
+			clip-rule="evenodd"
+			d="M419.31 37.9656C418.635 36.2071 419.007 33.547 420.087 32.3968C421.256 31.1537 424.259 30.8083 425.965 31.721C428.076 32.8507 428.455 36.8419 426.642 38.8528C424.909 40.7757 420.169 40.2024 419.31 37.9656ZM425.975 33.0156C427.143 34.1588 426.894 37.2098 425.544 38.3025C423.088 40.2909 420 38.7443 420 35.5253C420 32.3383 423.697 30.7853 425.975 33.0156ZM422.646 37.1133C422.136 38.5039 421.5 37.6022 421.5 35.4889C421.5 33.2601 421.515 33.2396 423.216 33.2396C424.521 33.2396 425.012 33.4915 425.266 34.2897C425.449 34.8671 425.357 35.5813 425.061 35.8772C424.678 36.2606 424.681 36.6045 425.073 37.0764C425.806 37.9601 424.47 37.9921 423.538 37.1133C423.008 36.6135 422.829 36.6135 422.646 37.1133ZM423.22 35.2389C423.649 35.2389 424 35.014 424 34.7391C424 34.4642 423.789 34.2393 423.53 34.2393C423.271 34.2393 422.92 34.4642 422.75 34.7391C422.58 35.014 422.792 35.2389 423.22 35.2389Z"
+			fill="white"
+		/>
+		<path
+			fill-rule="evenodd"
+			clip-rule="evenodd"
+			d="M435.875 143.974C434.431 144.168 430 144.5 430 144.5L430.321 142C430.798 142.356 432.872 141.69 434.668 141.512C438.329 141.148 438.5 141.033 438.5 138.937C438.5 137.028 438.369 136.991 434.5 137.786C434.036 137.881 433.62 137.965 433.246 138.04L433.242 138.041C431.21 138.451 430.321 138.5 430.321 138.5V136.5L431.698 136.992C433.101 136.809 435.207 136.42 436.375 136.128L438.5 135.599V106.917C438.5 88.5534 438.324 78.2352 438.011 78.2352C437.742 78.2352 431.375 78.4562 428 79V78L434.125 77.4334L438.5 76.6317V74.6843C438.5 73.4112 438.273 72.7369 437.845 72.7369C437.484 72.7369 433.333 72.9697 429.5 73.5L429 71.5L436 70L438.25 69.9758L438.386 50.4879L438.522 31L440.487 35.4196C441.567 37.8503 443.025 40.591 443.726 41.5092C444.609 42.6663 445 43.804 445 45.212V47.2449H447.172C448.747 47.2449 449.607 46.955 450.297 46.1922C450.821 45.6129 452.15 44.6572 453.25 44.0679C455.199 43.0237 455.836 42.9927 478.25 42.8493C499.828 42.7113 501.513 42.7643 505.5 43.7055C507.838 44.2573 510.987 44.9321 512.5 45.2055L512.899 45.2779C513.32 45.3541 513.721 45.427 514.123 45.4966C522.781 46.9953 531.698 46.9953 732.567 46.995L741.5 46.995H957.75L965.25 45.8448C969.375 45.212 974.475 44.2558 976.583 43.7205C980.105 42.8253 982.237 42.7473 1002.83 42.7573C1024.27 42.7678 1025.35 42.8123 1027.49 43.782C1028.72 44.3398 1030.44 45.347 1031.33 46.0203C1032.42 46.856 1033.58 47.2449 1034.97 47.2449H1037V45.2675C1037 43.9654 1037.44 42.6428 1038.29 41.3937C1039.42 39.7318 1040.16 38.2697 1042.68 32.7495C1043.12 31.7848 1043.28 36.1019 1043.38 51.6965L1043.51 71.8932L1051.63 71.5603C1061.28 71.1644 1065.71 70.3661 1071.04 68.0609L1075 66.3489V103.118V139.887L1072.77 142.027C1071.29 143.439 1069.64 144.437 1067.89 144.964C1064.73 145.92 1056.27 146.694 1048.88 146.706L1043.5 146.714V163.542C1043.5 172.797 1043.35 180.22 1043.17 180.037C1042.98 179.854 1042.09 177.985 1041.17 175.885C1040.26 173.784 1038.96 171.343 1038.29 170.462C1037.52 169.46 1037.01 168.033 1036.91 166.658C1036.76 164.65 1036.62 164.444 1035.31 164.294C1034.29 164.178 1033.1 164.664 1031.22 165.958C1026.56 169.168 1026.21 169.207 1002.45 169.207C981.418 169.207 980.979 169.186 975.964 167.972C970.062 166.543 961.797 165.379 953.25 164.771C946.753 164.309 536.654 164.505 528 164.974C522.262 165.285 512.907 166.398 509.138 167.216C507.824 167.501 505.288 168.079 503.5 168.501C500.776 169.144 496.645 169.243 478 169.112C455.896 168.958 455.735 168.949 453.423 167.8C452.144 167.163 450.702 166.207 450.22 165.675C449.61 165.001 448.685 164.708 447.172 164.708H445V166.728C445 168.079 444.658 169.114 443.968 169.852C443.4 170.459 442.317 172.306 441.56 173.955C441.257 174.615 440.985 175.225 440.74 175.774L440.739 175.777L440.738 175.778C439.893 177.674 439.372 178.842 439.051 178.789C438.537 178.705 438.536 175.486 438.536 167.109C438.535 165.592 438.535 163.905 438.532 162.037L438.5 143.621L435.875 143.974ZM496.914 51.1187L505.75 52.4933H740.75H975.75L983.5 51.2436C990.859 50.057 992.284 49.994 1011.75 49.994H1032.25L1034.06 53.6549C1035.06 55.6683 1036.14 58.5923 1036.47 60.1529C1036.9 62.2362 1037.01 74.8123 1036.9 107.476L1036.75 151.962L1035.43 155.343C1034.71 157.203 1033.59 159.511 1032.93 160.472L1031.75 162.218L1012 162.081C993.645 161.954 991.543 161.847 982.25 160.569L972.25 159.194L738.25 159.332L504.25 159.469L495.75 160.834C487.599 162.142 486.484 162.198 468.632 162.203L450.014 162.209L448.751 160.301C448.057 159.253 446.986 157.059 446.37 155.428L445.25 152.462L445.105 106.976C444.999 73.5871 445.115 60.8921 445.543 59.2411C445.863 58.004 446.91 55.3614 447.871 53.368L449.619 49.7441H468.849C487.352 49.7441 488.413 49.7961 496.914 51.1187ZM506.504 46.4621C504.719 46.0753 502.604 45.5309 501.805 45.252C500.082 44.6517 460.967 44.6582 457.398 45.2595C456.105 45.4775 454.536 46.0133 453.912 46.4502C452.869 47.1804 455.093 47.2419 481.263 47.2054C507.013 47.1694 509.438 47.0979 506.504 46.4621ZM441.983 43.2132C442.332 43.8655 442.495 64.3335 442.484 106.068C442.468 168.761 442.444 169.706 440.82 169.706C440.644 169.706 440.5 141.028 440.5 105.976C440.5 65.3722 440.675 42.2465 440.983 42.2465C441.248 42.2465 441.698 42.6813 441.983 43.2132ZM478.25 166.601C499.209 166.735 501.024 166.677 504.75 165.763L508.75 164.781L480.75 164.764C454.086 164.749 452.821 164.789 454.25 165.603C455.593 166.368 458.107 166.472 478.25 166.601ZM972.775 160.737C976.337 161.028 980.038 161.455 981 161.686C981.962 161.917 873.962 162.109 741 162.112C592.454 162.116 499.865 161.941 500.846 161.657C501.723 161.404 504.761 161.005 507.596 160.772C515.468 160.125 964.856 160.091 972.775 160.737ZM502 50.4274C505.946 51.2401 976.775 51.2336 980 50.4209C981.687 49.9955 921.829 49.8546 740.75 49.8571C559.546 49.8601 499.937 50.0025 502 50.4274ZM1027.25 46.0983L1029.25 47.2009L1000.75 47.1679C984.088 47.1489 973.081 46.9615 974.25 46.7166C975.35 46.4866 977.263 45.9638 978.5 45.5549C980.374 44.9356 984.469 44.8267 1003 44.9036C1024.92 44.9941 1025.28 45.0121 1027.25 46.0983ZM1002.75 166.679C1024.87 166.651 1027.86 166.435 1028.38 164.833C1028.54 164.338 1022.8 164.208 1000.79 164.208C985.506 164.208 973 164.379 973 164.588C973 164.797 974.743 165.359 976.875 165.838C980.25 166.595 983.587 166.704 1002.75 166.679ZM1072.5 137.68C1072.5 139.468 1069.93 141.779 1066.94 142.673C1064.41 143.432 1055.29 144.196 1048.62 144.206L1043.5 144.214V142.59V140.965L1052.12 140.953C1057.69 140.946 1061.73 140.7 1063.5 140.26C1066.42 139.536 1070.72 137.575 1071.24 136.726C1071.85 135.741 1072.5 136.232 1072.5 137.68ZM1067.24 75.9334C1071.27 74.9407 1072.5 74.076 1072.5 72.2256C1072.5 70.4881 1072.28 70.4641 1068.3 71.7707C1063.85 73.2312 1060.05 73.7506 1051.38 74.0835L1043.5 74.3859V76.0604V77.7353H1047.18C1052.93 77.7353 1063.94 76.7461 1067.24 75.9334ZM1072.5 104.153V132.57L1071.36 134.269C1068.69 138.229 1062.75 139.811 1051.35 139.604L1043.75 139.466L1043.62 109.42L1043.49 79.3748L1046.12 79.0594C1047.57 78.886 1051 78.74 1053.75 78.736C1059.31 78.727 1066.7 77.7258 1069.75 76.5662C1070.85 76.1483 1071.92 75.7904 1072.12 75.771C1072.33 75.752 1072.5 88.5234 1072.5 104.153ZM1040.25 169.207C1040.92 169.759 1041 162.657 1041 105.788C1041 70.5651 1040.94 41.7466 1040.87 41.7466C1040.8 41.7466 1040.46 42.1935 1040.12 42.7393C1039.63 43.5161 1039.5 57.3382 1039.5 106.158C1039.5 161.038 1039.59 168.66 1040.25 169.207ZM1037 160.209C1037 161.875 1036.83 162.209 1036 162.209C1035.45 162.209 1035 162.104 1035 161.976C1035 161.614 1036.67 158.21 1036.84 158.21C1036.93 158.21 1037 159.11 1037 160.209ZM1036.96 53.9928L1036.98 51.8684C1037 50.0705 1036.85 49.7441 1036 49.7441C1035.95 49.7441 1035.89 49.7439 1035.84 49.7437C1035.45 49.742 1035.21 49.7409 1035.12 49.8559C1034.98 50.0374 1035.22 50.5076 1035.83 51.7196L1035.98 52.0289L1036.96 53.9928ZM447 161.971C447 162.102 446.55 162.209 446 162.209C445.178 162.209 445.001 161.876 445.002 160.334L445.004 158.46L446.002 160.096C446.551 160.997 447 161.84 447 161.971ZM445 51.577C445 53.6394 445.048 53.6309 446.199 51.3686C446.998 49.7986 446.992 49.7441 446.013 49.7441C445.18 49.7441 445 50.0695 445 51.577Z"
+			fill="white"
+		/>
+
+		<foreignObject x="100" y="75" width="220" height="55">
+			<input
+				type="text"
+				name="Character Name"
+				id="char-name"
+				class="bg-transparent border-0 w-full h-full text-4xl"
+				bind:value={charName}
+			/>
+		</foreignObject>
+		<foreignObject x="80" y="135" width="160" height="160">
+			<span class="text-xs">CHARACTER NAME</span>
+		</foreignObject>
+
+		<line
+			x1="460"
+			y1="90"
+			x2="1020"
+			y2="90"
+			stroke="white"
+			stroke-opacity="0.7"
+		/>
+		<text x="465" y="102" fill="white" class="text-xs">CLASS & LEVEL</text>
+		<foreignObject x="465" y="70" width="220" height="55">
+			<Select
+				options={VALID_CLASSES}
+				bind:selected={selectedClass}
+				label="Select your Class"
+			/>
+		</foreignObject>
+
+		<text x="700" y="102" fill="white" class="text-xs">BACKGROUND</text>
+		<foreignObject x="700" y="70" width="220" height="55">
+			<Select
+				options={VALID_BACKGROUNDS}
+				bind:selected={selectedBackground}
+				label="Select your Background"
+			/>
+		</foreignObject>
+
+		<text x="900" y="102" fill="white" class="text-xs">PLAYER NAME</text>
+		<foreignObject x="900" y="70" width="220" height="55">
+			<input
+				class="border-0 placeholder:text-xs p-0 placeholder:text-white focus:outline-0 focus:shadow-none focus:ring-0"
+				type="text"
+				name="Player Name"
+				id=""
+				placeholder="Type your name"
+				bind:value={playerName}
+			/>
+		</foreignObject>
+		<line
+			x1="460"
+			y1="130"
+			x2="1020"
+			y2="130"
+			stroke="white"
+			stroke-opacity="0.7"
+		/>
+
+		<text x="465" y="144" fill="white" class="text-xs">RACE</text>
+		<foreignObject x="465" y="110" width="220" height="55">
+			<Select
+				options={VALID_RACES}
+				bind:selected={selectedRace}
+				label="Select your Race"
+			/>
+		</foreignObject>
+
+		<text x="700" y="144" fill="white" class="text-xs">ALIGMENT</text>
+		<foreignObject x="700" y="110" width="220" height="55">
+			<Select
+				options={VALID_ALIGNMENTS_2024}
+				bind:selected={selectedAlignment}
+				label="Select your Alignment"
+			/>
+		</foreignObject>
+
+		<text x="900" y="144" fill="white" class="text-xs">EXPERIENCE POINTS</text>
+		<foreignObject x="900" y="110" width="220" height="55">
+			<input
+				class="border-0 placeholder:text-xs p-0 placeholder:text-white focus:outline-0 focus:shadow-none focus:ring-0"
+				type="number"
+				name="Experience Points"
+				id=""
+				placeholder="Type your current exp"
+				bind:value={exp}
+			/>
+		</foreignObject>
+	</svg>
+</div>
+
+<style>
+	@import url("https://fonts.googleapis.com/css2?family=Lavishly+Yours&display=swap");
+
+	:global(#char-name) {
+		font-family: "Lavishly Yours", cursive;
+		font-weight: 400;
+		font-style: normal;
+		text-align: center;
+	}
+</style>
